moving the SOs and fixing rest to be non syncable
