+++ matchmaker/Assets/ServerObjects/Coin.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63355fe93e6f45448826e0e423b89a27
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
