+++ matchmaker/Assets/ServerObjects/Experiment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4276c8c5b94efe84b8ef4502d913e960
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
