+++ matchmaker/Assets/ServerObjects/Experiment/ExperimentRestSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 96da1fea09e16dd45ac15ae17362e4c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
